Time: 0 ms (100.00%) | Memory: 8.2 MB (22.98%) - LeetSync

+++ 50-powx-n/powx-n.cpp
@@ -14,7 +14,7 @@ public:
                 nn = nn / 2 ;   //   (x)^n  =  x*((x)^2)^(n-1)/2
             }
         }
-        if (n < 0) return (1.0)/(double)(ans) ;
+        if (n < 0) return 1.0/ans ;
         return (ans) ;
     }
 };
